System/hitbox/gurvan

+++ Libraries/Network/Packets/Structures/NewEntity.hpp
@@ -15,7 +15,7 @@
  * 
  */
 struct packet_new_entity {
-    int id; // id of new entity;
+    uint16_t id; // id of new entity;
     float positionX; // x position of new entity;
     float positionY; // y position of new entity;
     uint16_t direction; // direction of new entity;
